requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic_settings==2.6.1
supabase==2.11.0
pytest==8.3.4
factory-boy==3.3.1
pytest-asyncio==0.25.2
pytest-cov==6.0.0
pytest-mock==3.14.0
fastapi-users[sqlalchemy]
python-jose[cryptography]
passlib[bcrypt]
google-auth>=2.22.0
google-cloud-storage>=2.10.0